cobalt/kernel: pipeline: abstract context switching support
Implement an abstract API for the low-level context switching code,
moving the legacy open coded support to the I-pipe specific section
(e.g. fpu management, register file switching, root context tracking).
Dovetail provides built-in support for all these nitty-gritty details,
which we may tap into for the same purpose instead.

The changes have been introduced in a way which should allow for a
straightforward implementation of the new interface based on Dovetail
calls, simply by populating the interface services with such calls. In
other words, the final logic is already in place as a result of these
changes.

No functional change is introduced.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>