cobalt/clock: drop timer calibration
The calibrated timer setup time is currently accounted for in the
timer gravity triplet (.user, also used as default .irq latency),
exclusively. This forces in a dynamically calculated parameter with no
way to override it. Meanwhile, this value would be already included in
every gravity value which autotune may determine, so this setting is
pretty much redundant.

Drop nktimerlat and the requirement for the machine section to provide
a timer calibration handler, we don't need these.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>