cobalt/init: ipipe: remove clock frequency override
Regarding the frequency of the clock hardware, the kernel knows
better. This parameter dates back to the dark ages when the kernel
might not have detected variations in (per-CPU) clock frequency, which
is no more an issue nowadays.

Actually, we do want to rely on the non-trivial logic the kernel has
to figure out that value dynamically for us.

So let's drop the user override for this parameter.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>