cobalt/assert: pipeline: add TODO() marker
As its name suggests, TODO() can be added to the code in order to
highlight a place where code is known to be missing. Turning on
CONFIG_XENO_TODO reveals the places where such markers might linger by
triggering a build time assertion, so that they don't go unnoticed in
a comment.

We may drop this helper when the port to Dovetail is complete.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>