cobalt/sched: dovetail: disable separate per-IRQ statistics
The execution time of an IRQ handler should not be significant,
otherwise the implementation is broken in the first place. Too high an
interrupt rate is what can make things go wrong through, in which case
the information we would need to figure this out could be derived from
reading /proc/interrupts over Dovetail. Therefore, disable per-IRQ
stats when interfacing with Dovetail.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>